+++ R/project.R
@@ -1,3 +1,9 @@
+
+
+# Project class definition ------------------------------------------------
+
+
+
 #' Portable Encapsulated Project object
 #'
 #' Provides an in-memory representation and functions to access project
@@ -8,27 +14,55 @@
 #' @slot file character vector path to config file on disk.
 #' @slot samples a data table object holding the sample metadata
 #' @slot config a list object holding contents of the config file
+#' @slot sampleNameAttr a string indicating the sample attribute that is used
+#'  to index the sample table
+#' @slot subSampleNameAttr a string indicating the sample attribute that is used
+#'  to index the sample table
 #'
 #' @export
-setClass("Project",
-         slots = c(
-             file = "character",
-             samples = "data.frame",
-             config = "list"
-         ))
+setClass(
+    "Project",
+    slots = c(
+        file = "character",
+        samples = "data.frame",
+        config = "list",
+        sampleNameAttr = "character",
+        subSampleNameAttr = "character"
+    )
+)
+
+
+# Project class methods ---------------------------------------------------
 
 
 setMethod("initialize", "Project", function(.Object, ...) {
     .Object = methods::callNextMethod(.Object)  # calls generic initialize
     ellipsis <- list(...)
+    stIndex = ellipsis$sampleTableIndex
+    sstIndex = ellipsis$subSampleTableIndex
     if (!is.null(ellipsis$file)) {
-        # check if file path provided
-        .Object@file = .makeAbsPath(ellipsis$file, parent = path.expand(getwd()))
-        # instantiate config object and stick it in the config slot
-        .Object@config = Config(.Object@file, ellipsis$amendments)
-        .Object = .loadSampleAnnotation(.Object)
-        .Object = .modifySamples(.Object)
+        # check if 'file' path provided
+        if (.isCfg(ellipsis$file)) {
+            # provided 'file' seems to be a config
+            .Object@file = .makeAbsPath(ellipsis$file, parent = path.expand(getwd()))
+            # instantiate config object and stick it in the config slot
+            .Object@config = Config(.Object@file, ellipsis$amendments)
+            # determine the effective (sub)sample table indexes
+            .Object = .getTableIndexes(.Object, stIndex, sstIndex)
+            sampleTablePath = .getSampleTablePathFromConfig(config = config(.Object))
+            .Object@samples = .loadSampleAnnotation(sampleTablePath = sampleTablePath)
+            .Object = .modifySamples(.Object)
+        } else {
+            # provided 'file' seems to be a sample table
+            # determine the effective (sub)sample table indexes
+            .Object = .getTableIndexes(.Object, stIndex, sstIndex)
+            .Object@samples = .loadSampleAnnotation(sampleTablePath = ellipsis$file)
+            .Object = .autoMergeDuplicatedNames(.Object)
+        }
     }
+    # no 'file' provided, creating an empty object
+    # determine the effective (sub)sample table indexes
+    .Object = .getTableIndexes(.Object, stIndex, sstIndex)
     return(.Object)
 })
 
@@ -37,15 +71,28 @@ setMethod("initialize", "Project", function(.Object, ...) {
 #'
 #' This is a helper that creates the project with empty samples and config slots
 #'
-#' @param file a character with project configuration yaml file
-#' @param amendments a character with the amendments names to be activated
+#' @param file a string specifying a path to a project configuration YAML file
+#' @param amendments a string with the amendments names to be activated
+#' @param sampleTableIndex a string indicating the sample attribute that is used
+#'  to index the sample table
+#' @param subSampleTableIndex a string indicating the sample attribute that is used
+#'  to index the sample table
 #' @examples
 #' projectConfig = system.file("extdata", "example_peps-master",
 #' "example_amendments1", "project_config.yaml", package="pepr")
 #' p=Project(projectConfig)
 #' @export
-Project = function(file = NULL, amendments = NULL) {
-    methods::new("Project", file = file, amendments = amendments)
+Project = function(file = NULL,
+                   amendments = NULL,
+                   sampleTableIndex = NULL,
+                   subSampleTableIndex = NULL) {
+    methods::new(
+        "Project",
+        file = file,
+        amendments = amendments,
+        sampleTableIndex = sampleTableIndex,
+        subSampleTableIndex = subSampleTableIndex
+    )
 }
 
 
@@ -57,7 +104,7 @@ setMethod(
         cat("  file: ", object@file, fill = T)
         cat("  samples: ", NROW(object@samples), fill = T)
         if (length(object@config) != 0) {
-            .listAmendments(object@config, style="cat")
+            .listAmendments(object@config, style = "cat")
         }
         invisible(NULL)
     }
@@ -71,7 +118,7 @@ setMethod(
 #'
 #' @param object an object of \code{"\linkS4class{Project}"}
 #'
-#' @return project config 
+#' @return project config
 #'
 #' @examples
 #' projectConfig = system.file("extdata", "example_peps-master",
@@ -93,9 +140,9 @@ setMethod(
 )
 
 
-#' Perform all the sample addribute modifications
+#' Perform all the sample attribute modifications
 #'
-#' @param object an object of \code{"\linkS4class{Project}"} 
+#' @param object an object of \code{"\linkS4class{Project}"}
 #'
 #' @return modified Project object
 setGeneric(".modifySamples", function(object)
@@ -109,7 +156,9 @@ setMethod(
         object = .appendAttrs(object)
         object = .duplicateAttrs(object)
         object = .implyAttrs(object)
-        object = .mergeAttrs(object)
+        object = .autoMergeDuplicatedNames(object)
+        object = .mergeAttrs(object,
+                             .getSubSampleTablePathFromConfig(config(object)))
         object = .deriveAttrs(object)
         return(object)
     }
@@ -117,7 +166,7 @@ setMethod(
 
 
 #' Extract samples
-#' 
+#'
 #' This method extracts the samples
 #'
 #' @param .Object An object of Project class
@@ -145,19 +194,19 @@ setMethod(
     signature(.Object = "Project",
               sampleName = "character"),
     definition = function(.Object, sampleName) {
-        sampleNames = unlist(.Object@samples$sample_name)
+        sampleNames = unlist(.Object@samples[[.Object@sampleNameAttr]])
         rowNumber = which(sampleNames == sampleName)
         if (length(rowNumber) == 0)
             stop("Such sample name does not exist.")
-        result = sampleTable(.Object)[rowNumber, ]
+        result = sampleTable(.Object)[rowNumber,]
         return(result)
     }
 )
 
 
-#' Extract samples
-#' 
-#' This method extracts the samples
+#' Extract subsamples
+#'
+#' This method extracts the subsamples
 #'
 #' @param .Object An object of Project class
 #'
@@ -190,20 +239,20 @@ setMethod(
         subsampleName = "character"
     ),
     definition = function(.Object, sampleName, subsampleName) {
-        if (is.null(.Object@samples$subsample_name))
+        if (is.null(.Object@samples[[.Object@subSampleNameAttr]]))
             stop(
-                "There is no subsample_name attribute in the subannotation", 
-                " table, therefore this method cannot be called."
+                "There is no subsample_name attribute in the subsample table, ",
+                " therefore this method cannot be called."
             )
-        sampleNames = unlist(.Object@samples$sample_name)
+        sampleNames = unlist(.Object@samples[[.Object@sampleNameAttr]])
         rowNumber = which(sampleNames == sampleName)
         if (length(rowNumber) == 0)
             stop("Such sample name does not exist.")
-        subsampleNames = .Object@samples$subsample_name[[rowNumber]]
+        subsampleNames = .Object@samples[[.Object@subSampleNameAttr]][[rowNumber]]
         sampleNumber = which(subsampleNames == subsampleName)
         if (length(sampleNumber) == 0)
             stop("Such sample and sub sample name combination does not exist.")
-        result = .Object@samples[1,]
+        result = .Object@samples[1, ]
         for (iColumn in names(result)) {
             if (length(.Object@samples[[iColumn]][[rowNumber]]) > 1) {
                 result[[iColumn]] =
@@ -217,7 +266,7 @@ setMethod(
 )
 
 
-#' List amendments 
+#' List amendments
 #'
 #' Lists available amendments within a \code{"\linkS4class{Project}"} object.
 #'
@@ -225,7 +274,7 @@ setMethod(
 #'
 #' @param .Object an object of \code{"\linkS4class{Project}"}
 #' @return names of the available amendments
-#' @examples 
+#' @examples
 #' projectConfig = system.file("extdata",
 #' "example_peps-master",
 #' "example_amendments1",
@@ -243,7 +292,7 @@ setMethod(
     signature = signature(.Object = "Project"),
     definition = function(.Object) {
         config = config(.Object)
-        .listAmendments(cfg = config, style="message")
+        .listAmendments(cfg = config, style = "message")
     }
 )
 
@@ -259,8 +308,8 @@ setMethod(
 #'
 #' @param .Object an object of class \code{"\linkS4class{Project}"}
 #' @param amendments character with the amendment name
-#' 
-#' @examples 
+#'
+#' @examples
 #' projectConfig = system.file("extdata",
 #' "example_peps-master",
 #' "example_amendments1",
@@ -268,7 +317,7 @@ setMethod(
 #' package = "pepr")
 #' p = Project(file = projectConfig)
 #' availAmendments = listAmendments(p)
-#' activateAmendments(p,availAmendments[1])
+#' activateAmendments(p, availAmendments[1])
 #' @export
 setGeneric("activateAmendments", function(.Object, amendments)
     standardGeneric("activateAmendments"))
@@ -276,15 +325,16 @@ setGeneric("activateAmendments", function(.Object, amendments)
 #' @describeIn activateAmendments activate amendments in a \code{"\linkS4class{Project}"} object
 setMethod(
     f = "activateAmendments",
-    signature = signature(.Object="Project", amendments="character"),
+    signature = signature(.Object = "Project", amendments = "character"),
     definition = function(.Object, amendments) {
         .Object@config = .applyAmendments(.Object@config, amendments)
         .Object@config = makeSectionsAbsolute(
-            .Object@config, 
-            c(CFG_SAMPLE_TABLE_KEY, CFG_SUBSAMPLE_TABLE_KEY), 
+            .Object@config,
+            c(CFG_SAMPLE_TABLE_KEY, CFG_SUBSAMPLE_TABLE_KEY),
             .Object@file
         )
-        .Object = .loadSampleAnnotation(.Object)
+        sampleTablePath = .getSampleTablePathFromConfig(config = config(.Object))
+        .Object@samples = .loadSampleAnnotation(sampleTablePath = sampleTablePath)
         .Object = .modifySamples(.Object)
         return(.Object)
     }
@@ -318,7 +368,7 @@ setMethod(
     }
 )
 
-# sample modifiers --------------------------------------------------------
+# Sample modifiers --------------------------------------------------------
 
 #' Remove attributes across all the samples
 #'
@@ -326,15 +376,17 @@ setMethod(
 #'
 #' @return an object of \code{"\linkS4class{Project}"}
 .removeAttrs <- function(.Object) {
-    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object))) return(.Object)
+    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object)))
+        return(.Object)
     modifiers = config(.Object)[[CFG_S_MODIFIERS_KEY]]
-    if (!CFG_REMOVE_KEY %in% names(modifiers)) return(.Object)
+    if (!CFG_REMOVE_KEY %in% names(modifiers))
+        return(.Object)
     toRemove = modifiers[[CFG_REMOVE_KEY]]
     if (!is.null(toRemove)) {
         # get a copy of samples to get the dimensions
         for (rem in toRemove) {
-            if(rem %in% colnames(sampleTable(.Object))) {
-                .Object@samples[,rem] = NULL
+            if (rem %in% colnames(sampleTable(.Object))) {
+                .Object@samples[, rem] = NULL
             }
         }
     }
@@ -344,13 +396,15 @@ setMethod(
 
 #' Append constant attributes across all the samples
 #'
-#' @param .Object an object of \code{\link{Project-class}} 
+#' @param .Object an object of \code{\link{Project-class}}
 #'
-#' @return an object of \code{\link{Project-class}} 
+#' @return an object of \code{\link{Project-class}}
 .appendAttrs <- function(.Object) {
-    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object))) return(.Object)
+    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object)))
+        return(.Object)
     modifiers = config(.Object)[[CFG_S_MODIFIERS_KEY]]
-    if (!CFG_APPEND_KEY %in% names(modifiers)) return(.Object)
+    if (!CFG_APPEND_KEY %in% names(modifiers))
+        return(.Object)
     constants = modifiers[[CFG_APPEND_KEY]]
     if (is.list(constants)) {
         # get names
@@ -358,11 +412,13 @@ setMethod(
         # get a copy of samples to get the dimensions
         colLen = dim(sampleTable(.Object))[1]
         for (iConst in seq_along(constants)) {
-            # create a one column data.table and appand it to the current one
-            if(!constantsNames[iConst] %in% colnames(sampleTable(.Object))) {
-                tempDT = data.table::data.table(
-                    matrix(matrix(NA, ncol=1, nrow=colLen)))
-                tempDT[, 1] = list(list(constants[[iConst]]))
+            # create a one column data.table and append it to the current one
+            if (!constantsNames[iConst] %in% colnames(sampleTable(.Object))) {
+                tempDT = data.table::data.table(matrix(matrix(
+                    list(constants[[iConst]]),
+                    ncol = 1,
+                    nrow = colLen
+                )))
                 names(tempDT) = constantsNames[iConst]
                 .Object@samples = cbind(.Object@samples, tempDT)
             }
@@ -377,12 +433,14 @@ setMethod(
 #'
 #' @return an object of \code{"\linkS4class{Project}"}
 .duplicateAttrs <- function(.Object) {
-    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object))) return(.Object)
+    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object)))
+        return(.Object)
     modifiers = config(.Object)[[CFG_S_MODIFIERS_KEY]]
-    if (!CFG_DUPLICATE_KEY %in% names(modifiers)) return(.Object)
+    if (!CFG_DUPLICATE_KEY %in% names(modifiers))
+        return(.Object)
     duplicated = modifiers[[CFG_DUPLICATE_KEY]]
-    for(oriAttrName in names(duplicated)){
-        .Object@samples[,duplicated[[oriAttrName]]] = .Object@samples[,oriAttrName]
+    for (oriAttrName in names(duplicated)) {
+        .Object@samples[, duplicated[[oriAttrName]]] = .Object@samples[, oriAttrName]
     }
     return(.Object)
 }
@@ -391,36 +449,44 @@ setMethod(
 #'
 #' @param .Object an object of \code{"\linkS4class{Project}"}
 #'
-#' @return an object of \code{"\linkS4class{Project}"} 
+#' @return an object of \code{"\linkS4class{Project}"}
 .implyAttrs = function(.Object) {
-    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object))) return(.Object)
+    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object)))
+        return(.Object)
     modifiers = config(.Object)[[CFG_S_MODIFIERS_KEY]]
-    if (!CFG_IMPLY_KEY %in% names(modifiers)) return(.Object)
+    if (!CFG_IMPLY_KEY %in% names(modifiers))
+        return(.Object)
     implications = modifiers[[CFG_IMPLY_KEY]]
     for (implication in implications) {
-        if (!(CFG_IMPLY_IF_KEY %in% names(implication) && CFG_IMPLY_THEN_KEY %in% names(implication)))
+        if (!(
+            CFG_IMPLY_IF_KEY %in% names(implication) &&
+            CFG_IMPLY_THEN_KEY %in% names(implication)
+        ))
             stop(CFG_IMPLY_KEY, " section is not formatted properly")
         implierAttrs = names(implication[[CFG_IMPLY_IF_KEY]])
         implierVals = implication[[CFG_IMPLY_IF_KEY]]
         impliedAttrs = names(implication[[CFG_IMPLY_THEN_KEY]])
         impliedVals = as.character(implication[[CFG_IMPLY_THEN_KEY]])
         attrs = colnames(.Object@samples)
-        if (!all(implierAttrs %in% attrs)) next
+        if (!all(implierAttrs %in% attrs))
+            next
         allHitIds = list()
         for (i in seq_along(implierAttrs)) {
             hitIds = list()
             implierStrings = as.character(unlist(implierVals[i]))
-            for(j in seq_along(implierStrings)){
-                hitIds[[j]] = which(.Object@samples[,implierAttrs[i]] == implierStrings[j])    
+            for (j in seq_along(implierStrings)) {
+                hitIds[[j]] = which(.Object@samples[, implierAttrs[i]] == implierStrings[j])
             }
             allHitIds[[i]] = Reduce(union, hitIds)
-            if (length(allHitIds[[i]]) < 1) break
+            if (length(allHitIds[[i]]) < 1)
+                break
         }
         qualIds = Reduce(intersect, allHitIds)
-        if (length(qualIds) < 1) next
-        for (i in seq_along(impliedAttrs)){
-            if (!impliedAttrs[i] %in% attrs) 
-                .Object@samples[,impliedAttrs[i]] = ""
+        if (length(qualIds) < 1)
+            next
+        for (i in seq_along(impliedAttrs)) {
+            if (!impliedAttrs[i] %in% attrs)
+                .Object@samples[, impliedAttrs[i]] = ""
             .Object@samples[qualIds, impliedAttrs[i]] = impliedVals[i]
         }
     }
@@ -434,24 +500,33 @@ setMethod(
 #'
 #' @return an object of \code{"\linkS4class{Project}"}
 .deriveAttrs = function(.Object) {
-    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object))) return(.Object)
+    if (!CFG_S_MODIFIERS_KEY %in% names(config(.Object)))
+        return(.Object)
     parentDir = dirname(.Object@file)
     modifiers = config(.Object)[[CFG_S_MODIFIERS_KEY]]
-    if (!CFG_DERIVE_KEY %in% names(modifiers)) return(.Object)
+    if (!CFG_DERIVE_KEY %in% names(modifiers))
+        return(.Object)
     derivations = modifiers[[CFG_DERIVE_KEY]]
-    if (!all(c(CFG_DERIVE_ATTRS_KEY, CFG_DERIVE_SOURCES_KEY) 
+    if (!all(c(CFG_DERIVE_ATTRS_KEY, CFG_DERIVE_SOURCES_KEY)
              %in% names(derivations)))
         stop(CFG_DERIVE_KEY, " section is not formatted properly")
     for (derivedAttr in derivations[[CFG_DERIVE_ATTRS_KEY]]) {
-        derivedSamplesVals = .Object@samples[,derivedAttr]
-        for (derivedSource in names(derivations[[CFG_DERIVE_SOURCES_KEY]])){
+        if (!derivedAttr %in% colnames(.Object@samples))
+            stop(paste(
+                "Failed to derive. Sample attribute not found:",
+                derivedAttr
+            ))
+        derivedSamplesVals = .Object@samples[, derivedAttr]
+        for (derivedSource in names(derivations[[CFG_DERIVE_SOURCES_KEY]])) {
             hitIds = which(derivedSamplesVals == derivedSource)
-            if (length(hitIds) < 1) next
-            for (hitId in hitIds){
+            if (length(hitIds) < 1)
+                next
+            for (hitId in hitIds) {
                 rgx = derivations[[CFG_DERIVE_SOURCES_KEY]][[derivedSource]]
-                res = .matchesAndRegexes(.strformat(
-                    rgx, as.list(sampleTable(.Object)[hitId,]), parentDir))  
-                .Object@samples[hitId,derivedAttr] = list(unique(unlist(res)))
+                res = .matchesAndRegexes(.strformat(rgx, as.list(sampleTable(
+                    .Object
+                )[hitId, ]), parentDir))
+                .Object@samples[hitId, derivedAttr] = list(unique(unlist(res)))
             }
         }
     }
@@ -459,23 +534,18 @@ setMethod(
 }
 
 
-#' Read sample annotation from disk
+#' Read sample table from disk
 #'
-#' @param .Object an object of \code{"\linkS4class{Project}"}
+#' @param sampleTablePath a character string indicating a path to the sample table
 #'
-#' @return an object of \code{"\linkS4class{Project}"}
-.loadSampleAnnotation = function(.Object) {
-    cfg = config(.Object)
-    if (!CFG_SAMPLE_TABLE_KEY %in% names(cfg)) return(.Object) 
-    sampleAnnotationPath = cfg[[CFG_SAMPLE_TABLE_KEY]]
-    if(.safeFileExists(sampleAnnotationPath)){
-        samples = data.table::fread(sampleAnnotationPath)
+#' @return an data.frame with samples; one sample per row
+.loadSampleAnnotation = function(sampleTablePath) {
+    if (.safeFileExists(sampleTablePath)) {
+        samples = data.table::fread(sampleTablePath)
     } else{
-        warning("The sample_table does not exist: ", sampleAnnotationPath)
-        return(.Object)
+        warning("The sample table does not exist: ", sampleTablePath)
+        samples = data.frame()
     }
-    .Object@samples = samples
-    return(.Object)
 }
 
 
@@ -487,11 +557,11 @@ setMethod(
 #' @return an object of \code{"\linkS4class{Project}"}
 .loadSubsampleAnnotation = function(.Object, path) {
     if (.safeFileExists(path)) {
-        samplesSubannotation = data.table::fread(path)
+        subsamplesTable = data.table::fread(path)
     } else{
-        samplesSubannotation = data.table::data.table()
+        subsamplesTable = data.table::data.table()
     }
-    subNames = unique(samplesSubannotation$sample_name)
+    subNames = unique(subsamplesTable[[.Object@sampleNameAttr]])
     samples = sampleTable(.Object)
     samples = .listifyDF(samples)
     rowNum = nrow(samples)
@@ -499,9 +569,9 @@ setMethod(
     # into the samples data.table as a column. This way the "cells"
     # in the samples table can consist of multiple elements
     for (iName in subNames) {
-        whichNames = which(samplesSubannotation$sample_name == iName)
-        subTable = samplesSubannotation[whichNames,]
-        dropCol = which(names(samplesSubannotation[whichNames,]) == "sample_name")
+        whichNames = which(subsamplesTable[[.Object@sampleNameAttr]] == iName)
+        subTable = subsamplesTable[whichNames, ]
+        dropCol = which(names(subsamplesTable[whichNames, ]) == .Object@sampleNameAttr)
         subTable = subset(subTable, select = -dropCol)
         colList = vector("list", rowNum)
         for (iColumn in seq_len(ncol(subTable))) {
@@ -514,8 +584,8 @@ setMethod(
             }
             # The column exists
             whichColSamples = which(names(samples) == colName)
-            whichRowSamples = which(samples$sample_name == iName)
-            if(length(whichRowSamples) < 1){
+            whichRowSamples = which(samples[[.Object@sampleNameAttr]] == iName)
+            if (length(whichRowSamples) < 1) {
                 warning("No samples named '", iName, "'")
             } else {
                 # Inserting element(s) into the list
@@ -535,14 +605,93 @@ setMethod(
 #' Merge samples defined in sample table with ones in subsample table(s)
 #'
 #' @param .Object an object of \code{"\linkS4class{Project}"}
+#' @param subsampleAannotationPaths a vector of strings specifying the paths to sample
 #'
 #' @return an object of \code{"\linkS4class{Project}"}
-.mergeAttrs = function(.Object){
-    cfg = config(.Object)
-    if (!CFG_SUBSAMPLE_TABLE_KEY %in% names(cfg)) return(.Object)
-    sampleSubannotationPath = cfg[[CFG_SUBSAMPLE_TABLE_KEY]]
-    for(p in sampleSubannotationPath){
-        .Object = .loadSubsampleAnnotation(.Object, p)    
+.mergeAttrs = function(.Object, subsampleAannotationPaths) {
+    if (is.null(subsampleAannotationPaths))
+        return(.Object)
+    for (p in subsampleAannotationPaths) {
+        .Object = .loadSubsampleAnnotation(.Object, p)
     }
+    return(.Object)
+}
+
+
+#' Merge samples with identical names
+#'
+#' If sample table specifies samples with non-unique names, try to merge these samples
+#'
+#' @param .Object an object of \code{"\linkS4class{Project}"}
+#'
+#' @return an object of \code{"\linkS4class{Project}"}
+.autoMergeDuplicatedNames = function(.Object) {
+    s = sampleTable(.Object)
+    sampleNames = s[[.Object@sampleNameAttr]]
+    dups = duplicated(sampleNames)
+    if (!any(dups))
+        return(.Object)
+    duplicatedSampleNames = sampleNames[which(dups)]
+    if (!is.null(.getSubSampleTablePathFromConfig(config(.Object))))
+        stop(
+            paste0(
+                "Duplicated sample names were found (",
+                duplicatedSampleNames,
+                ") and subsample_table is specified in the config. ",
+                "You may use either auto-merging or subsample_table-based merging."
+            )
+        )
+    rowsWithDuplicates = which(s[[.Object@sampleNameAttr]] %in% duplicatedSampleNames)
+    dupRows = s[rowsWithDuplicates,]
+    combinedList = list()
+    # create a list of list that combined rows with duplicated sample names
+    for (duplicatedSampleName in duplicatedSampleNames) {
+        toCombine = s[which(sampleNames == duplicatedSampleName),]
+        combinedList[[duplicatedSampleName]] = apply(toCombine, 2, function(x) {
+            return(list(unique(x)))
+        })
+    }
+    # remove rows that include duplicates
+    s = s[-rowsWithDuplicates,]
+    # insert the combined rows
+    for (combinedRow in combinedList) {
+        s = rbind(s, combinedRow)
+    }
+    .Object@samples = s
+    return(.Object)
+}
+
+#' Set table indexes
+#'
+#' Get sample and subsample table indexes and save as a slot on the Project object
+#'
+#' This is the (sub)sample table index selection priority order:
+#' \enumerate{
+#'   \item Project constructor specified
+#'   \item Config specified
+#'   \item Deafult value
+#' }
+#'
+#' @param .Object an object of \code{"\linkS4class{Project}"}
+#' @param stIndex character string indicating a constructor-specified sample table index
+#' @param sstIndex character string indicating a constructor-specified subsample table index
+.getTableIndexes <- function(.Object, stIndex, sstIndex) {
+    .getIndexVal <- function(spec, config, default, key) {
+        if (!is.null(spec))
+            return(spec)
+        if (length(config))
+            return(ifelse(is.null(config[[key]]), default, config[[key]]))
+        return(default)
+    }
+    
+    .Object@sampleNameAttr = .getIndexVal(stIndex,
+                                          config(.Object),
+                                          SAMPLE_NAME_ATTR,
+                                          CFG_SAMPLE_TABLE_INDEX_KEY)
+    .Object@subSampleNameAttr = .getIndexVal(sstIndex,
+                                             config(.Object),
+                                             SUBSAMPLE_NAME_ATTR,
+                                             CFG_SUBSAMPLE_TABLE_INDEX_KEY)
+    
     return(.Object)
 }